requirements.txt
requests 
python-dotenv 
faiss-cpu 
pydantic 
fastapi 
streamlit 
sentence-transformers 
uvicorn 
rag 
